update/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fonttools==4.14.0
idna==2.10
pathspec==0.8.0
regex==2020.7.14
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
